backend/requirements.txt
pyserial>=3.5
python-dotenv>=1.0.0
watchdog>=2.1.0
# Note: python3-pyudev should be installed via apt-get
# sudo apt-get install python3-pyudev 